Extracted Articles class to separate file.
